+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 * Added `WrappedPointer`
 * Added `NFData` and `Generic` instances for `Comp`
 * Added `Prod`
+* `Data.Extensible.Class` now exports `compareMembership`
 
 0.4.2
 -------------------------------------------------